Bug #3017821 - Add support for libproxy (felipec).
"libproxy is a library that provides automatic proxy configuration management, it's usage greatly simplifies proxy configuration in XChat. For example, it's able to detect whether GNOME or KDE is running, and fetch the configuration from there, as well as environment variables, DConf, and other sources can be configured as modules.
In order to maintain the current proxy configurations, nothing will change unless the user selects the "Auto" proxy type."


git-svn-id: https://svn.code.sf.net/p/xchat/svn@1489 893a96be-7f27-4fdf-9d1e-6aeec9d3cce1